Tidied test cases.
